+++ java/src/main/java/io/monohull/controller/ConfigController.java
@@ -64,6 +64,7 @@ public class ConfigController {
         entity.setDbVendor(req.dbVendor());
         entity.setDbName(req.dbName());
         entity.setDbContainerPort(req.dbContainerPort());
+        entity.setDbCommand(req.dbCommand());
         entity.setHostVolumePath(req.hostVolumePath());
         entity.setDbVolumeName(req.dbVolumeName());
         entity.setWorkspacePath(req.workspacePath());
@@ -99,6 +100,7 @@ public class ConfigController {
         entity.setDbVendor(req.dbVendor());
         entity.setDbName(req.dbName());
         entity.setDbContainerPort(req.dbContainerPort());
+        entity.setDbCommand(req.dbCommand());
         entity.setHostVolumePath(req.hostVolumePath());
         entity.setDbVolumeName(req.dbVolumeName());
         entity.setWorkspacePath(req.workspacePath());
@@ -189,6 +191,7 @@ public class ConfigController {
         config.setAppHttpsPort(req.appHttpsPort());
         config.setDbPort(req.dbPort());
         config.setDbPassword(req.dbPassword());
+        config.setDbCommand(req.dbCommand());
         config.setDbExtraEnv(req.dbExtraEnv());
         config.setDbExtraBinds(req.dbExtraBinds());
         config.setAppExtraEnv(req.appExtraEnv());
@@ -204,7 +207,7 @@ public class ConfigController {
         PipelineDefinitionEntity pipeline = e.getPipelineDefinition();
         return new ImageConfigResponse(e.getId(), e.getClient(), e.getProject(),
             e.getMaximoVersion(), e.getAppImage(), e.getDbImage(), e.getAdmImage(),
-            e.getDbVendor(), e.getDbName(), e.getDbContainerPort(),
+            e.getDbVendor(), e.getDbName(), e.getDbContainerPort(), e.getDbCommand(),
             e.getHostVolumePath(), e.getDbVolumeName(),
             e.getWorkspacePath(),
             e.getAppHttpPort(), e.getAppHttpsPort(), e.getDbPort(),
@@ -229,7 +232,7 @@ public class ConfigController {
             config.getId(), config.getHostVolumePath(),
             config.getDbVolumeName(), config.isStaticPorts(),
             config.getAppHttpPort(), config.getAppHttpsPort(), config.getDbPort(),
-            config.getDbPassword(),
+            config.getDbPassword(), config.getDbCommand(),
             config.getDbExtraEnv(), config.getDbExtraBinds(),
             config.getAppExtraEnv(), config.getAppExtraBinds(),
             config.getAdmExtraEnv(), config.getAdmExtraBinds(),