+++ java/src/main/java/io/monohull/dto/ConfigUpdateRequest.java
@@ -10,6 +10,7 @@ public record ConfigUpdateRequest(
     Integer appHttpsPort,
     Integer dbPort,
     String dbPassword,
+    String dbCommand,
     List<ExtraEnvVar> dbExtraEnv,
     List<ExtraBind> dbExtraBinds,
     List<ExtraEnvVar> appExtraEnv,