+++ java/src/main/java/io/monohull/dto/ImageConfigBundle.java
@@ -39,6 +39,7 @@ public record ImageConfigBundle(
         String dbVendor,
         String dbName,
         Integer dbContainerPort,
+        String dbCommand,
         String hostVolumePath,
         String dbVolumeName,
         String workspacePath,