+++ java/src/main/java/io/monohull/dto/ImageConfigResponse.java
@@ -14,6 +14,7 @@ public record ImageConfigResponse(
     String dbVendor,
     String dbName,
     Integer dbContainerPort,
+    String dbCommand,
     String hostVolumePath,
     String dbVolumeName,
     String workspacePath,